+++ packages/annotation/src/filters/BlurFilter.tsx
@@ -1,0 +1,9 @@
+import { BlurFilterProps } from './types'
+
+export const BlurFilter = ({ id, blurStdDeviation = 5 }: BlurFilterProps) => {
+    return (
+        <filter id={id}>
+            <feGaussianBlur stdDeviation={blurStdDeviation} />
+        </filter>
+    )
+}